lufthansa-travel-buddy/pyproject.toml
[tool.poetry]
name = "lufthansa-travel-buddy"
version = "0.1.0"
description = "Proactive disruption manager for Lufthansa business travelers - Hamburg Hackathon 2026"
authors = ["Your Team <team@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "^1.0.0"
google-genai = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"

[tool.poetry.group.deployment.dependencies]
google-cloud-aiplatform = "^1.50.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
